requirements.txt
Flask==3.0.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.2
gunicorn==21.2.0
flask-cors==4.0.0
requests==2.31.0
web3==6.11.3
Pillow==10.1.0
